+++ internal/plugin/service/mysql/user/zz_resource.go
@@ -116,6 +116,6 @@ func resourceSchema(ctx context.Context) schema.Schema {
 			},
 		},
 		Blocks:              map[string]schema.Block{"timeouts": legacytimeouts.BlockAll(ctx)},
-		MarkdownDescription: "Creates and manages an Aiven for MySQL® service user.",
+		MarkdownDescription: "Creates and manages an Aiven for MySQL® service user. If this resource is missing (e.g., after a service power off), it will be removed from the state and a new create plan will be generated.",
 	}
 }